fix the things up deply another dockerfile and add the api fixes add model support
