+++ apps/api/wink-bm25-text-search.d.ts
@@ -1,0 +1,2 @@
+// wink-bm25-text-search.d.ts
+declare module "wink-bm25-text-search";
