+++ apps/api/wink-nlp-utils.d.ts
@@ -1,0 +1,2 @@
+// wink-nlp-utils.d.ts
+declare module "wink-nlp-utils";
